+++ old.c
@@ -6,14 +6,14 @@
 /*   By: frapp <fabi@student.42.fr>                 +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/23 23:23:56 by frapp             #+#    #+#             */
-/*   Updated: 2024/05/26 03:09:12 by frapp            ###   ########.fr       */
+/*   Updated: 2024/06/16 06:45:13 by frapp            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "includes/cub3D.h"
 #include "MLX42/include/MLX42/MLX42.h"
 
-void	scale_triangle_3d(t_triangle *tri_a, float scalar)
+void	scale_triangle_3d(t_triangle *tri_a, double scalar)
 {
 	for (int i = 0; i < 3; i++)
 	{
@@ -21,7 +21,7 @@ void	scale_triangle_3d(t_triangle *tri_a, float scalar)
 	}
 }
 
-void	scale_mesh_3d(t_mesh *mesh, float scalar)
+void	scale_mesh_3d(t_mesh *mesh, double scalar)
 {
 	for (int i = 0; i < mesh->count; i++)
 	{
